+++ src/main/java/org/owasp/webgoat/lessons/deserialization/InsecureDeserializationTask.java
@@ -1,31 +1,8 @@
-/*
- * This file is part of WebGoat, an Open Web Application Security Project utility. For details, please see http://www.owasp.org/
- *
- * Copyright (c) 2002 - 2019 Bruce Mayhew
- *
- * This program is free software; you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation; either version 2 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program; if
- * not, write to the Free Software Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA
- * 02111-1307, USA.
- *
- * Getting Source ==============
- *
- * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
- */
-
 package org.owasp.webgoat.lessons.deserialization;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InvalidClassException;
-import java.io.ObjectInputStream;
 import java.util.Base64;
 import org.dummy.insecure.framework.VulnerableTaskHolder;
 import org.owasp.webgoat.container.assignments.AssignmentEndpoint;
@@ -35,37 +12,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
-import java.io.InputStream;
-import java.util.Set;
-import java.util.HashSet;
-
-
-public class SecureObjectInputStream extends ObjectInputStream {
-    private static final Set<String> allowedClasses = new HashSet<>();
-
-    static {
-        allowedClasses.add("org.dummy.insecure.framework.VulnerableTaskHolder");
-        allowedClasses.add("java.lang.String");
-        allowedClasses.add("java.time.LocalDateTime");
-        allowedClasses.add("java.time.LocalDate");
-        allowedClasses.add("java.time.LocalTime");
-    }
-
-    public SecureObjectInputStream(InputStream in) throws IOException {
-        super(in);
-    }
-
-    @Override
-    protected Class<?> resolveClass(ObjectStreamClass desc) throws IOException, ClassNotFoundException {
-        String className = desc.getName();
-        if (!allowedClasses.contains(className)) {
-            throw new InvalidClassException("Clase no autorizada: " + className);
-        }
-        return super.resolveClass(desc);
-    }
-}
-
-
 
 @RestController
 @AssignmentHints({
@@ -75,42 +21,42 @@ public class SecureObjectInputStream extends ObjectInputStream {
 })
 public class InsecureDeserializationTask extends AssignmentEndpoint {
 
-  @PostMapping("/InsecureDeserialization/task")
-  @ResponseBody
-  public AttackResult completed(@RequestParam String token) throws IOException {
-    String b64token;
-    long before;
-    long after;
-    int delay;
+    @PostMapping("/InsecureDeserialization/task")
+    @ResponseBody
+    public AttackResult completed(@RequestParam String token) throws IOException {
+        String b64token;
+        long before;
+        long after;
+        int delay;
 
-    b64token = token.replace('-', '+').replace('_', '/');
+        b64token = token.replace('-', '+').replace('_', '/');
 
-    try (ObjectInputStream ois =
-        new SecureObjectInputStream(new ByteArrayInputStream(Base64.getDecoder().decode(b64token)))) {
-      before = System.currentTimeMillis();
-      Object o = ois.readObject();
-      if (!(o instanceof VulnerableTaskHolder)) {
-        if (o instanceof String) {
-          return failed(this).feedback("insecure-deserialization.stringobject").build();
+        try (SecureObjectInputStream ois =
+                new SecureObjectInputStream(new ByteArrayInputStream(Base64.getDecoder().decode(b64token)))) {
+            before = System.currentTimeMillis();
+            Object o = ois.readObject();
+            if (!(o instanceof VulnerableTaskHolder)) {
+                if (o instanceof String) {
+                    return failed(this).feedback("insecure-deserialization.stringobject").build();
+                }
+                return failed(this).feedback("insecure-deserialization.wrongobject").build();
+            }
+            after = System.currentTimeMillis();
+        } catch (InvalidClassException e) {
+            return failed(this).feedback("insecure-deserialization.invalidversion").build();
+        } catch (IllegalArgumentException e) {
+            return failed(this).feedback("insecure-deserialization.expired").build();
+        } catch (Exception e) {
+            return failed(this).feedback("insecure-deserialization.invalidversion").build();
         }
-        return failed(this).feedback("insecure-deserialization.wrongobject").build();
-      }
-      after = System.currentTimeMillis();
-    } catch (InvalidClassException e) {
-      return failed(this).feedback("insecure-deserialization.invalidversion").build();
-    } catch (IllegalArgumentException e) {
-      return failed(this).feedback("insecure-deserialization.expired").build();
-    } catch (Exception e) {
-      return failed(this).feedback("insecure-deserialization.invalidversion").build();
-    }
 
-    delay = (int) (after - before);
-    if (delay > 7000) {
-      return failed(this).build();
+        delay = (int) (after - before);
+        if (delay > 7000) {
+            return failed(this).build();
+        }
+        if (delay < 3000) {
+            return failed(this).build();
+        }
+        return success(this).build();
     }
-    if (delay < 3000) {
-      return failed(this).build();
-    }
-    return success(this).build();
-  }
 }
